[ETCM-697] Cleanup EthTransactionResponse usage
