+++ src/main/scala/io/iohk/ethereum/jsonrpc/TransactionResponse.scala
@@ -31,23 +31,6 @@ final case class TransactionResponse(
     input: ByteString
 ) extends BaseTransactionResponse
 
-final case class EthTransactionResponse(
-    hash: ByteString,
-    nonce: BigInt,
-    blockHash: Option[ByteString],
-    blockNumber: Option[BigInt],
-    transactionIndex: Option[BigInt],
-    from: Option[ByteString],
-    to: Option[ByteString],
-    value: BigInt,
-    gasPrice: BigInt,
-    gas: BigInt,
-    input: ByteString,
-    r: ByteString,
-    s: ByteString,
-    v: ByteString
-) extends BaseTransactionResponse
-
 final case class TransactionData(
     stx: SignedTransaction,
     blockHeader: Option[BlockHeader] = None,
@@ -78,32 +61,4 @@ object TransactionResponse {
       input = stx.tx.payload
     )
 
-}
-
-object EthTransactionResponse {
-
-  def apply(tx: TransactionData): EthTransactionResponse =
-    EthTransactionResponse(tx.stx, tx.blockHeader, tx.transactionIndex)
-
-  def apply(
-      stx: SignedTransaction,
-      blockHeader: Option[BlockHeader] = None,
-      transactionIndex: Option[Int] = None
-  ): EthTransactionResponse =
-    EthTransactionResponse(
-      hash = stx.hash,
-      nonce = stx.tx.nonce,
-      blockHash = blockHeader.map(_.hash),
-      blockNumber = blockHeader.map(_.number),
-      transactionIndex = transactionIndex.map(txIndex => BigInt(txIndex)),
-      from = SignedTransaction.getSender(stx).map(_.bytes),
-      to = stx.tx.receivingAddress.map(_.bytes),
-      value = stx.tx.value,
-      gasPrice = stx.tx.gasPrice,
-      gas = stx.tx.gasLimit,
-      input = stx.tx.payload,
-      r = UInt256(stx.signature.r).bytes,
-      s = UInt256(stx.signature.s).bytes,
-      v = ByteString(stx.signature.v)
-    )
 }
