+++ src/main/scala/io/iohk/ethereum/testmode/TestEthBlockServiceWrapper.scala
@@ -1,8 +1,8 @@
 package io.iohk.ethereum.testmode
 
-import io.iohk.ethereum.domain.{Block, Blockchain, UInt256}
+import io.iohk.ethereum.domain.{Block, BlockHeader, Blockchain, SignedTransaction, UInt256}
 import io.iohk.ethereum.jsonrpc.EthBlocksService.{BlockByBlockHashResponse, BlockByNumberResponse}
-import io.iohk.ethereum.jsonrpc.{BaseBlockResponse, BaseTransactionResponse, EthBlocksService, EthTransactionResponse, ServiceResponse, TransactionData}
+import io.iohk.ethereum.jsonrpc.{BaseBlockResponse, BaseTransactionResponse, EthBlocksService, ServiceResponse, TransactionData}
 import io.iohk.ethereum.ledger.Ledger
 import io.iohk.ethereum.utils.Logger
 import io.iohk.ethereum.consensus.Consensus
@@ -106,3 +106,48 @@ case class EthBlockResponse(
     transactions: Either[Seq[ByteString], Seq[BaseTransactionResponse]],
     uncles: Seq[ByteString]
 ) extends BaseBlockResponse
+
+final case class EthTransactionResponse(
+    hash: ByteString,
+    nonce: BigInt,
+    blockHash: Option[ByteString],
+    blockNumber: Option[BigInt],
+    transactionIndex: Option[BigInt],
+    from: Option[ByteString],
+    to: Option[ByteString],
+    value: BigInt,
+    gasPrice: BigInt,
+    gas: BigInt,
+    input: ByteString,
+    r: ByteString,
+    s: ByteString,
+    v: ByteString
+) extends BaseTransactionResponse
+
+object EthTransactionResponse {
+
+  def apply(tx: TransactionData): EthTransactionResponse =
+    EthTransactionResponse(tx.stx, tx.blockHeader, tx.transactionIndex)
+
+  def apply(
+      stx: SignedTransaction,
+      blockHeader: Option[BlockHeader] = None,
+      transactionIndex: Option[Int] = None
+   ): EthTransactionResponse =
+    EthTransactionResponse(
+      hash = stx.hash,
+      nonce = stx.tx.nonce,
+      blockHash = blockHeader.map(_.hash),
+      blockNumber = blockHeader.map(_.number),
+      transactionIndex = transactionIndex.map(txIndex => BigInt(txIndex)),
+      from = SignedTransaction.getSender(stx).map(_.bytes),
+      to = stx.tx.receivingAddress.map(_.bytes),
+      value = stx.tx.value,
+      gasPrice = stx.tx.gasPrice,
+      gas = stx.tx.gasLimit,
+      input = stx.tx.payload,
+      r = UInt256(stx.signature.r).bytes,
+      s = UInt256(stx.signature.s).bytes,
+      v = ByteString(stx.signature.v)
+    )
+}
